Added user management by buildings and signup links with permission checks

+++ admin_edit_user.php
@@ -55,18 +55,47 @@ if(!empty($_POST['action'])){
 			$error['name'] = "Invalid name";
 		}
 
-		if (trim($_POST['apartment']) && !preg_match('/Z[0-9]\.B[0-9]\.[0-9]{3}/i',$_POST['apartment'])){
+		$group_for_new_user = null;
+
+		if (trim($_POST['apartment']) && !preg_match('/Z([0-9])\.B([0-9])\.[0-9]{3}/i',$_POST['apartment'], $m)){
 			$error['apartment'] = "Use format ZX.BY.NNN for apartment";
 		}
 
+		if(!_check_permission('super_admin')) {
+			$_POST['default_garage'] = 'z'.$m[1];
+
+			if(!_check_admin_building_permission($_SESSION['id'], $_POST['apartment'])){
+				$error['apartment'] = "You cannot use this apartment, check your permissions";
+			}
+		}
+
+		$group_for_new_user = 'Z'.$m[1].'.B'.$m[2];
+
 		$edit_user['name'] = htmlspecialchars($_POST['name']);
 		$edit_user['enabled'] = !empty($_POST['enabled']);
+		$edit_user['can_create_guests'] = !empty($_POST['can_create_guests']);
+		$edit_user['has_camera_access'] = !empty($_POST['has_camera_access']);
 
-		if (empty($_POST['groups'])){
-			$error['groups'] = "Choose at least one group";
+		if(_check_permission('super_admin')) {
+			if (empty($_POST['groups'])) {
+				$error['groups'] = "Choose at least one group";
+			}
 		}
 
 		$user_groups = $_POST['groups'];
+
+		if(!$_POST['id']){
+			$group_for_new_user2 = $group_for_new_user;
+			$group_for_new_user = $db->fetch('SELECT id FROM `groups` WHERE name=?', $group_for_new_user);
+
+			if(empty($group_for_new_user)){
+				$error['groups'] = "No applicable group. Contact super admin".$group_for_new_user2;
+			}
+		}
+
+		if($_POST['id'] && !_check_admin_permission_for_user($_SESSION['id'], $_POST['id'])){
+			$error['permission'] = "You are not authorized to edit this user.";
+		}
 
 		if(empty($error)){
 			$smarty->assign('success', 'Successfully saved');
@@ -74,8 +103,11 @@ if(!empty($_POST['action'])){
 			// save all except pwd
 			if(!$_POST['id']){
 				// create new user
-				$db->query('INSERT INTO users SET username=?, password=?,name=?,created=NOW(),enabled=#,has_camera_access=#,can_create_guests=#,apartment=?,default_garage=?,email=?,phone=?', $_POST['username'],PasswordTools::getHashedPassword($_POST['password']), $_POST['name'], $_POST['enabled'] ? 1 : 0, $_POST['has_camera_access'] ? 1 : 0,$_POST['can_create_guests'] ? 1 : 0, $_POST['apartment'],$_POST['default_garage'], $_POST['email'], $_POST['phone']);
+				$db->query('INSERT INTO users SET username=?, password=?,name=?,created=NOW(),enabled=#,has_camera_access=#,can_create_guests=#,apartment=?,default_garage=?,email=?,phone=?,created_by=#', $_POST['username'],PasswordTools::getHashedPassword($_POST['password']), $_POST['name'], $_POST['enabled'] ? 1 : 0, $_POST['has_camera_access'] ? 1 : 0,$_POST['can_create_guests'] ? 1 : 0, $_POST['apartment'],$_POST['default_garage'], $_POST['email'], $_POST['phone'], $_SESSION['id']);
 				$_GET['id'] = $_POST['id'] = $db->lastinsertid();
+
+				$db->query('INSERT INTO user_group SET user_id=#, group_id=#', $_POST['id'], $group_for_new_user);
+
 			}else{
 				// update user
 				$db->query('UPDATE users SET username=?, name=?,enabled=#,has_camera_access=#,can_create_guests=#,apartment=?,default_garage=?,email=?,phone=? WHERE id=#', $_POST['username'], $_POST['name'], $_POST['enabled'] ? 1 : 0,  $_POST['has_camera_access'] ? 1 : 0,$_POST['can_create_guests'] ? 1 : 0, $_POST['apartment'],$_POST['default_garage'], $_POST['email'], $_POST['phone'], $_POST['id']);
@@ -86,10 +118,12 @@ if(!empty($_POST['action'])){
 				}
 			}
 
-			// save groups
-			$db->query('DELETE FROM user_group WHERE user_id=#', $_POST['id']);
-			foreach ($_POST['groups'] as $g){
-				$db->query('INSERT INTO user_group SET user_id=#, group_id=#', $_POST['id'], $g);
+			if(_check_permission('super_admin')) {
+				// save groups
+				$db->query('DELETE FROM user_group WHERE user_id=#', $_POST['id']);
+				foreach ($_POST['groups'] as $g) {
+					$db->query('INSERT INTO user_group SET user_id=#, group_id=#', $_POST['id'], $g);
+				}
 			}
 
 		}else{
@@ -98,10 +132,18 @@ if(!empty($_POST['action'])){
 		}
 
 	}elseif($_POST['action'] == 'delete'){
-		$db->query('DELETE FROM users WHERE id=#', $_POST['id']);
+		// check if permissions and also, cannot delete yourself
+		if(_check_admin_permission_for_user($_SESSION['id'], $_POST['id']) && $_SESSION['id'] != $_POST['id']) {
+			$db->query('DELETE FROM users WHERE id=#', $_POST['id']);
+		}
 		header('Location: admin_users.php');
 		exit;
 	}
+}
+
+if($_GET['id'] && !_check_admin_permission_for_user($_SESSION['id'], $_GET['id'])){
+	header('Location: admin_users.php');
+	exit;
 }
 
 $user = _get_user_array();
